feat: add vim-like keybinds
Add new keybinds for selecting and going back:
"l" does the same as "enter" when choosing a project or task to view.
"h" goes back to the previous view, from show -> task,
from task -> project.

I thought that when you're viewing the project list view,
we don't want to quit by pressing "h".
"Esc" or "q" is better suited for that.