feat: read task labels when creating the initial project list
Add a new helper function GetAllLabels that is called when creating the initial project list model. This adds unique labels across all tasks in all projects to the projectListModel. The labels are mapped to their occurence, thus making it possible to show labels with higher frequency at the top of the multi-select list that will be added to the task form.