fix: tell jj where to fetch and rebase from (#42)
This commit tries to fix a problem that would arise when a remote receives changes while yatto is running and want's to commit and push a change itself.

Furthermore, if a backend command fails it's stdout/stderr output will be shown in a error message. Before that error messages would only show the return code of the failed command.